next-ops-pointers-recursively
