+++ src/utils/index.js
@@ -83,3 +83,81 @@ module.exports.stagesByTypes = (data) => {
   // return stages
   return [firstStage, secondStage, thirdStage]
 }
+
+/**
+ * loadConfNextOperations recursively loop over
+ * the object and formats each instance of next operation
+ * or on error to conf format.
+ *
+ * @param {object} data - base data object before format
+ * @param {object} workflows - all workflows
+ * @param {object} operations - all operations
+ */
+module.exports.loadConfNextOperations = (data, workflows, operations) => {
+  Object.keys(data).forEach(key => {
+    // check if recursion reaches next operation
+    // or on error instances and format it
+    if (key === "next_operation" || key === "on_error") {
+      // take deep copy of next operation
+      const nextOp = { ...(data[key] || {}) }
+
+      // reset next operations
+      data[key] = {}
+
+      // add next operation
+      data[key].selector = Object.entries(nextOp).map(([wid, oid]) => {
+        return {
+          workflow: workflows[wid].key,
+          operation: operations[oid].key
+        }
+      })
+    }
+
+    // go nested
+    if (typeof data[key] === "object") {
+      this.loadConfNextOperations(data[key], workflows, operations)
+    }
+  })
+}
+
+/**
+ * loadDocNextOperations recursively loop over
+ * the object and formats each instance of next operation
+ * or on error to doc format.
+ *
+ * @param {object} data - base data object before format
+ * @param {object} workflows - all workflows
+ */
+module.exports.loadDocNextOperations = (data, workflows) => {
+  Object.keys(data).forEach(key => {
+    // check if recursion reaches next operation
+    // or on error instances and format it
+    if (key === "next_operation" || key === "on_error") {
+      // destract next operations and on errors
+      const nextOpSelector = (data[key] || {}).selector || []
+
+      // reset next operation
+      data[key] = {}
+
+      // attach next operation
+      nextOpSelector.forEach(select => {
+        // find workflow id by key
+        const workflowId = Object.keys(workflows || {}).find(
+          workflowId => workflows[workflowId].key === select.workflow
+        )
+
+        if (workflowId) {
+          // set next operation by workflow id
+          data[key][workflowId] = select.operation
+        } else {
+          throw new Error(`You try to upload a operation with unknown workflow with key = ${select.workflow}. please make sure you upload also the workflow that belongs to this operation.`)
+        }
+      })
+    }
+
+    // go nested
+    if (typeof data[key] === "object") {
+      this.loadDocNextOperations(data[key], workflows)
+    }
+  })
+}
